+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @deltakit/react
 
+## 0.2.0
+
+### Minor Changes
+
+- Add fromAgnoAgents converter for Agno agent framework support
+
+### Patch Changes
+
+- Updated dependencies
+  - @deltakit/core@0.2.0
+
 ## 0.1.4
 
 ### Patch Changes